Organised the VizPsyche src folder
